HashObject: change strategy on hash map key-value mapping
originally this should be one big json string blob (as implemented by simple-json column type), but now we should use seperated key column as primary key instead. this is because in some database that doesn't support a long json string nor supporting json in general this could be a problem (namely oracle). JSON spec also clarifies that key entry should be string, and we can preemptively do a little optimization as well.